requirements.txt
# Dependencies for FastAPI endpoints (Railway deployment)
fastapi==0.104.1
pydantic>=2.8.0
python-dotenv==1.0.0
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.11
pandas>=2.0.0
